mac os x port
Relies on gnu coreutils (brew install coreutils) which are
installed with the g prefix.

pipe2 is a Linux specific syscall

ucontext on mac os x is a different structure. I took the easy
way out and disabled it. If needed we will need to add the
proper dump_registers implementation.

Fixes #2