adds stf test for ipv4 option processing (#10)
This test is an initial attempt to Vladimir's IPv4 options handling
(https://barefootnetworks.atlassian.net/wiki/display/~vgurevich/IPv4+Options+Handling). However, it's not yet worth even half a point, as it only
forwards to CPU, not even copy to CPU, which would imply a clone.

The test is not currently compiling.