fixes paths for new repo in driver script and enables it in runtest
Use the new configuration for the assembler and linker in the driver
script. Also fixes the output options for assembler and linker so that
all the generated files are in the same output directory. Added a
tofino-psa target that is currently not implemented.

Updated the runtest to invoke the driver rather than each binary. For
gdb, we still need to directly invoke the backend binary, so some of
the arguments are set differently depending on the debug/no-debug
mode. The .tfa file is now generated in the same directory too, so
there is no need for runtest to move it when running STF tests.