Warm up simple_switch libtool script
The bootstrap script builds bmv2 simple_switch. Although it also installs it, the bmv2 stf script uses the local version, which happens to actually be a libtool wrapper script. This script takes a long time to start the actual binary the first time (needs to locate and load the shared libraries), which can cause a timeout in the stf script. This seems like an acceptable hack to solve this problem.