adds cpplint to assembler sources
For now there are 3449 errors, so the cpplint rules in the assembler
are just for completion -- serious cleanup will need to happen later.

Also, a fair number of headers are common between bf-asm and
p4c. Given that they live in the same repo, it will make sense to
remove the duplication.