select a subset of the tests to run in Travis
The current selection is the tests in the p4c repo (both p4_14 and
p4_16) and stateful. The choice is based on the assumption that the
public repo has more diverse tests. Jenkins should be testing the full
suite (testdata and p4-tests) nightly.

Also updates the reference point for p4c to ensure that test labels
are properly attached.