adds reasons for failures for Tofino xfails
With this PR there are two ways in which one can tag tests as xfail:

1. Recommended: use the `p4c_add_xfail_reason` macro and provide a
regular expression that matches the expected error. When the test
passes or moves to a different error, ctest will report it as a
failure.

2. Add the tests to TOFINO_XFAIL_TESTS
This is the way we've been doing it until now. It is still used
for tests that pass the compilation step and fail in the model.

All the compiler failures for the Tofino backend are characterized
now. WIP to characterize the p14_to_16 conversions and bmv2 tests.

This PR requires https://github.com/p4lang/p4c/pull/827 to be merged.