fixes null pointer dereferences
field_defuse: phv.field(hdr_ids.first) returns null in some instances
(inline-bmv2.p4). In these instances, we just return, not try to read
the field.

instruction_selection and input_xbar seem to have issues with
operand[4] for a hash primitive in flowlet_switching (both P4_14 and
P4_16 versions). It seems to me that this is a program error, however,
the compiler should not crash. I am currently just printing a warning.