Debian packaging with cpack (#432)
* packaging and versioning

* compile JBay optionally in assembler

Introduces HAVE_JBAY as a configuration time macro that can be enabled
using -DENABLE_JBAY=ON when running cmake to optionally compile JBay
into the assembler.

Also removes the installation of chip.schema which are not needed for
generating binaries (just for register templates) and removes it as a
parameter that is passed to walle by the p4c driver.

* Add missing HAVE_JBAY include guard

Change CPACK_STRIP_FILES ON to CPACK_STRIP_FILES TRUE

Update p4c submodule ref point

Add LICENSE and README

* bundles chip.schema with walle executable

Using pyinstaller's add-data feature to bundle the schema files with
the walle executable in the lib/<chip>/chip.schema.

We add a --target <chip_name> argument to walle (and tflink) to allow
us to retrieve the schema.

To find the file, we check whether we're running as a bundle (see
https://pyinstaller.readthedocs.io/en/stable/runtime-information.html#run-time-information).
Pyinstaller sets the frozen attribute, and the sys._MEIPASS variable
to the root directory of the bundle. We then set the appropriate path
for the schema using the chip argument.

walle still accepts a --schema argument so that an explicit file can
be passed. It is overwritten when bundled, but can be used when
running out of the python source -- so we can merge these changes in
all the other numerous copies of walle.

tflink no longer needs to search the path for the schema, because in
our setup it always uses the bundled version. p4c.tofino.cfg is now
invoking tflink with a --target argument. I also removed the find_file
from tflink, it should always be invoked with an explicit walle
executable (whether python or bundled), and changed both places where
it was called: p4c driver and bf-asm/test/runtests. Couldn't find
other instances, but we should keep an eye for them.

* Removing README from dockerignore

So that Cpack does not complain.