fixes STF lexing context for keywords
Adds a new state to the STF lexer to lex tokens as keywords or as
plain IDs when not at the beginning of a command.

Also adds testing of STF generation. Creates a new category of tests,
TEST_STF2PTF_GEN to run the stf2ptf_pd generator over all existing
STF tests, as well as a make rule `check-stf2ptf-gen` to exercise
these tests.

Therefore, to pass all the tests, had to add a few more features:
- tcam_2bit_mode support in the lexer and the parser. PTF doesn't know
  what to do with them so the statements are ignored.
- lexing support for header stacks (e.g. data[0]). The parser will
  escape the brackets and generate data_0_ as a name. Don't know yet
  how PD is handling header stacks (or P4Runtime for that matter!)
- fixes encoding of odd-length packets for PD generation