fixes for stf2ptf translation
The main change is a reimplementation of the STF2PTF packet checking
using the PTF packet comparison support. The previous packet checking
was based on an earlier version of the framework. Further refinement
is possible by taking advantage of the PTF Mask object.

Also fixes an number of other issues:
- proper escaping of stack variables. STF defines its own syntax for
  stack variables (using name$index), and we also added support for
  name[index]. P4Runtime uses the latter and PD escapes the [index]
  with _index_.
- handling of fully qualified names for fields and standard
  metadata. P4runtime requires fully qualified names, while STF does
  not. For standard metadata we chose to have the translator add the
  "standard_metadata." qualifier, and for the couple of tests from p4c
  we adjusted the names (therefore the p4c refpoint update).

With these changes, a number of STF tests now complete successfully.
Tofino xfails have been adjusted to reflect it.