add debugging information and cleanup of intermediate files
Adds -g to the assembler and linker to decide whether to save
intermediate files.

If not enabled, then the assembler deletes the input assembly file.
We chose to use -g since the driver stops before invoking the assembler
if it was invoked with -S.

If not enabled, the linker deletes the p4i file, the assembler log file.
And the linker removes all cfg.json files regardless.

Adds -g to the compiler backend for future compatibility on generating
debug information/resource usage.