Add ability to use custome PTF test directory
As opposed to default behavior, which is to look for the existence of
the `<P4_source_dir>/<P4_source_name>.ptf` directory and use it. This
default behavior is inconvenient for programs like tor.p4, which are in
a submodule, as the PTF tests are expected to be updated more frequently
than the P4 source. In some cases we may also not have control over the
submodule containing the P4 source.