fixes the reading of counters from p4runtime (#601)
* fixes the reading of counters from p4runtime

A read request returns a list of counters, and one has to find
the counter matching the id and the index/match entry.
I'm still guessing that for direct counters we can just compare
table entries, would be nice to get an STF test for direct
counters too!

* implement proper comparison for indirect counters

This implements programming indirect counters for p4runtime and
comparison of table entries using SerializeToString.

It is also a first attempt of reconciling the semantics of
packet/expect/wait with P4runtime. This is not a full solution,
most of the existing "Expected packet" failures are likely due
to wrong handling of the semantics.

* refactored processing of STF statements into STFRunner

Refactored the code shared by stf2ptf.py, stf2ptf_pd, and now stf_test
into an abstract class (STFRunner) that implements the semantics of
executing STF. It defines a set of abstract methods to process
functionality for each statement (or groups of statements). Derived
classes must implement these methods.

The reason for this design is that it was getting ridiculous to debug
using the model and switchd. Replicating the code to just print the
sequence of statements and their grouping led to the problem that
changes need to get back. Meanwhile, stf2ptf_pd was rotting. And
writing the same method 3 times is not something that I care to do.

This patch brings stf2ptf.py, stf_test and stf2ptf_pd on top of the
STFRunner class. stf_test, simply prints the sequence of statements
that get executed, making it a lot easier to debug.

* remove the pairing of packets and wait for all of them at the end

Implements the semantics of STF packet/expect by issueing all
packet sends and then checking that the expected packets were
received only at the end. PTF buffers the packets for us in order.

However, this implies that we can't pair packets and expects, so
we take advantage of the Mask object to handle ternary data in
expect statements.

* use ignore_extra_bytes on ptf masks

* treat header name references in STF as header.valid