more lenient on which pyinstaller versions are broken (#664)
It does seem there was a period when pip installed a dev version
of pyinstaller that is broken -- 3.3-dev0 or such. Since, they
moved to a fully supported 3.3 version that seems to work.

We are still installing by default 3.2.1 as our dependency, since
that has been thoroughly tested by both Brig and Glass, but let
3.3 go through given that a number of Jenkins VMs are installed
with it. Eventually we should relax the 3.2.1 pinning.