Refactor translation passes (#675)
Previously, the translation passes relies on the visitor framework with a pass_manager for each type of externs to translate. The implementation has quite a few code duplications and inefficiency. Further, because the intermediate P4 program is not a valid P4 program, the typeMap and refMap are useless.

This refactor adopts how the translation is done in the P4-14 to 16 flow, by using a P4-16 program structure to keep all the intermediate translation states. These states are maps from the original IR nodes to the translated IR nodes. The AnalyzeV1modelProgram and ConstructSymbolTable pass initially inspect the p4 program and populate the translation maps with the original IR nodes; the GenerateTofinoProgram pass invokes various converters to process the entries in the translation maps and generates a valid p4 program.

A PathExpressionCloner is used in various places to make copies of IR::Path if the IR node is shared between ingress / egress parser or deparser.