stateful tables should not be implemented as meters
The assembler considered the stateful tables as meters.
This implied they should have the fields required by meters,
such as meter_granularity. This PR adds another category of
attached tables, 'stateful', and the places the stateful tables
in it. It also renames the 'stateful_alu_width' attribute to
'alu_width' per the necessary schema.

The PR also introduces a 'is_indirect' method for Table::Call,
so that we replace the obscure tests: args.size() >= 1, or args.size() > 0.

There is only an initial set of checks, and we print a warning that
there needs to be more. More testing is also required.

BRIG-233  #assign #start-work #fixed #fixed