adds support for testing code generation only (#821)
* adds support for testing code generation only

Adds a new tag 'codegen' that can be used to add tests that simply
check that the correct assembly has been generated. This is accomplished
by adding a new test driver (codegen_runner.sh) that runs the compiler
and if successful, prints out the assembly output. The test writer then
uses ctest regular expressions to filter this output and decide whether
the test is successful or not.

The initial set of tests are the p4_16 programs. This will need to be
adjusted for the proper tests.

* add the cmake file

* fix xfail order

* merge Xfails and Tests for all other test suites but Tofino and JBay

* add example macros