adds support for mandatory tests
Tests that are mandatory are listed in a variable named
"TAG"_MUST_PASS_TESTS where TAG is TOFINO or JBAY. The
p4c_add_xfail_reason macro checks this list and will not add the test
as an XFAIL. For now, we just print a warning message. That will turn
into a FATAL_ERROR so that you should not be able to configure if
you've added the wrong test to xfail (failing early is preferred).

And that's it really, the list of mandatory pass tests has similar
syntax with the other tests.