adds support for @pragma command_line
Glass supports a subset of command line options to be passed as
pragmas. We subset that set and add support for only two of the options
`--no-dead-code-elimination` and `--placement`.

The `--placement` option in glass takes parameters, while p4c doesn't,
so we trim it. The other options for placement do not make sense.