removes device specific info from testExtraArgs
TestUtils sets testExtraArgs globally to define the testing
environment. However, because this is global and tests can be added
for either Tofino or JBay, we need to avoid hardcoding the device into
the testExtraArgs. We add the device directly when adding tests using
the p4c_add_test macro.

Also adds switch as an example of CodeGen test and comments out the other
placeholder examples.