Created a reference for p4-tests and deleted jenkins folder (#977)
* Created a reference for p4-tests and deleted jenkins folder

The tests in the programs directory have their includes at the same
level with the p4. We add a macro that allows us to eliminate the
tofino.p4 file. In fact, that should be included from the compiler,
but that is a separate PR.

In addition there are several tests that now succeed, several new
tests that hit new bugs, and several new tests that touch
unimplemented features (e.g. dynamic hashes).