Update Dockerfile.dockerhub to reduce build time
Building the docker image takes too long at the moment and dockerhub
cancels the job. The docker image is used for hardware regressions which
means we are running the tests with an out-of-date image.

This commit switches to a DEBUG build which should speed up compilation
and disables the bmv2 backend as well as the gtests. Disabling the
gtests required updating the p4c submodule refpoint.