fixes to build on ubuntu 14.04 with gcc-4.9
OS dependencies that need to be manually installed include:
- boost 1.58.0
- cmake 3.5.2

In addition gcc-4.9 can not convert std::stringstream to a string,
so we call the str() when used in logging.

And for some reason that I'm too tired to debug, gtest doesn't link
boost::optional with bitranges. So I disabled those tests when
compiling with gcc-4.9 (and not clang).