Fix direct counter read in stf2ptf
After P4Runtime update to the DirectCounterEntry message (according to
the PSA spec a match-table can have at most one direct resource of each
type, so there is no need for the counter_id field, given that the
table_entry field already includes the table_id).

This commit also includes a priority-related field: when creating the
STFNamedEntry we need to store the actual P4Runtime priority, not the
original STF priority.