Send several packets in tor.p4 Packet IO PTF tests
The punt table in tor.p4 is gated by a meter. There used to be a PI
issue where the default meter spec configuration was wrong and every
packet but the first one was dropped. This wasn't detected right away
because the tests would only send one packet, so we have increased the
number of packets sent.