Improve how PHV allocation uses action analysis (#1144)
* Part 1/2: Update ActionAnalysis PHV interface

Given a proposed container packing, action analysis (via the can_pack method)
currently returns no/yes/conditional constraints.  PHV allocation, however,
treats conditional constraints as "yes" without actually using them.

This is the first of two commits fixing how PHV allocation uses action
analysis.  It does two things:

1. lifts analysis from fields to slices
2. changes the can_pack interface to return either boost::none, meaning no
   packing is possible, or a set of conditional constraints on source operand
   packing, with the empty set representing no extra constraints

However, although the functionality for (2) is in place, PHV allocation does
not yet properly handle conditional constraints.  (That will be the second of
these two commits.)  For now, conditional constraints are replaced with
boost::none.

* Part 2/2: Use conditional action constraints

This commit updates PHV allocation to respect conditional constraints produced
by action analysis.

KNOWN BUG: Recursive invoking tryAllocSliceList with conditional constraints
needs to map conditional constraints to slice lists, otherwise placement will
always fail because partial slice lists with exact containers requirements will
not fill entire containers.

* Paper over $stkvalid (again)

For some reason, the $stkvalid alias was only set up when a push operation was
present.  This commit forces the alias to always be created, and adds special
handling in use/def to tie uses/defs of $valid to $stkvalid.

* Overlay mutually exclusive metadata

Some sets of tables are mutually exclusive, i.e. for all packets either one set
or the other (but not both) are applied.  The same is true for some sets of
actions.  Metadata used exclusively in one set is then mutually exclusive with
metadata used exclusively in another, and can be overlaid.

This is similar to how headers are mutually exclusive based on the control-flow
graph of the parser: Some parser states are mutually exclusive.

This commit generalizes the parser live range analysis to detect mutually
exclusive fields by looking at mutually exclusive control-flow paths.

It also relaxes a check in validate_allocation to only check fields that may be
live at the same time.

* Add extraction constraints

If two slices are both extracted, then they can only be packed together if they
are extracted together (in the right alignment).

This commit introduces a conservative fix: If N slices are extracted, then they
can be packed together if they are adjacent.  Adjacency is determined by slice
lists.

This is conservative in that it rejects some valid packings, eg. f1, f2, and f3
8b fields extracted together (in that order).  Allocating [f1, ..., f3, ...] to
a 32b container, leaving holes (or allocating non-extracted fields) at bits
[15:8] and [31:24], is valid but will be rejected.

This only affects fields that are extracted but not deparsed.  (Deparsed fields
have stronger constraints.)  At present, only intrinsic metadata and bridged
metadata fields are extracted but not deparsed.

* Refactor PHV::Field::foreach_byte

This commit slightly changes the semantics of PHV::Field::foreach_byte to
invoke a user-supplied function on the segment of each alloc_slice in each
container byte, from field LSB to MSB.  Different slices of the same field
allocated to different slices of the same byte of the same container will cause
the function to be invoked on different slices of the same byte.

This fixes a bug in PHV::Field::foreach_byte, wherein it failed when different
adjacent slices of a field were allocated to non-adjacent slices of the same
container.

Along the way, this commit also adds comments and tests and refactors the code
to rely more heavily on the bitrange library.

* Update JBay XFAILs.

* Temporary HACK for bridged metadata

Force sub-32b bridged fields to be placed in the smallest container possible.
This is only necessary until the bridged metadata fix is in.

* Add switch_dc_basic to XFAILs (temporarily)

* Disable overlaying for POV bits because they are unconditionally read by the deparser.