Parser generate save and match instead of using match on input buffer sugar. (#1197)
* Compiler generate save and select on Match Registers.

This commit change the way how parser handle select. Previously, parser relies
on the syntax sugar that compiler only need to specify `match [offset]` for
selects and matches and asm will generate the save to register and select on
registers between states. However, this is not what it should be, since it's
only a syntax sugar created only to manually write bfa file easier and it make
supporting select on fields extracted in previous states impossible, which leads
to the lookahead issue.

Now, parser generate save and select on match register after resolve all
computed values, before phv. Then when splitting the states, split saves
to the state when it is possible to be inside the input buffer. Match registers
are allocated using following algorithm:

1. On postorder of every state, add selects to unresolved select of this state.
   add itself to unprocessed state.
2. Forall transitions from this state,
    a. Get all unresolved select of the next state, shift them by the `shift` value.
    b. Allocate match register to them, if they are extracted in this state, otherwise,
       add it to unresolved selets of this state. Note that there are several requirements
       on register allocation, e.g. if the select has already been allocated with a specific
       register, we must use that register on all the other branches.
    c. Mark corresponding select that it is using this register.
3. In the end_apply, if there are unprocess state that has unresolved selets (TNA case),

This commit also fixes some bugs, like ignoring casting in select expression,
boundary check in splitting state, hacks for non byte-align constants in select.

* optimize match register allocation

* add id to match register and use ordered_set when it needs to be iterated