fixes for brig-482 and brig-479 (#1250)
Resubmit and Mirror handling in extract_maupipe consists of two steps.
The first step traverse the frontend IR and collects the emitted field
lists. The second step modifies the backend IR to replace the
placeholder parser states with states that extracts the resubmit and
mirror fields. As we refactored extract_maupipe to be two pass managers
to be applied to the frontend and backend IR respectively, we should
also split the resubmit and mirror field list handling code to two
separate passes.