misc packaging, build, and test fixes (#1236)
* misc packaging, build, and test fixes

A variety of misc fixes including:
- complete the list of package dependecies when creating the DEB and RPM packages
- check for rapidjson install and skip installation if present
- switch tests get higher priority, so that they overlap with additional tests

* removed RPM packages

I tried verifying the RPM package we're building on Fedora and it
is a total disaster:
- we link versions of the Ubuntu libraries that are not available on Fedora
- we install in /usr/local which is not allowed
- we do not respect many of their packaging guidelines for license, file layout,
permissions. See https://fedoraproject.org/wiki/Packaging:Guidelines?rd=Packaging/Guidelines

So I parked it on a different branch. In the meantime I did update install_OS_deps.sh
to build on Fedora, but I'll keep it with the other branch.

* update p4c refpoint

* add tar.bz2 packaging