adds a loop around reading counters in p4runtime tests
Since PTF does not have a mechanism to ensure completion of packet
sends in the model, we have been getting intermittent errors on
reading counters. We have implemented a 1s wait, but there are
instances where that is not sufficient.

Therefore we add a while loop to repeatedly read the counter to
ensure that we have enough time to receive the packets. We sleep for
1s between iterations, and we iterate 20 odd times.