Handle bridged metadata fields extracted in Phase0 (#1292)
Right now, bridged metadata fields are packed into a header in the midend,
which induces alignment requirements.  If Phase0 fields are bridged, then they
need to be extracted in the ingress parser, which induces additional alignment
constraints.

The long-term fix is to account for Phase0 alignment constraints when
synthesizing/reordering "flexible" bridged metadata headers in the backend.
See BRIG-470.

The short-term fix, implemented here, is to prevent bridged metadata fields
that are also extracted in Phase0 from being packed together, which takes
advantage of the fact that, for each individual field, the packing layout
between Phase0 and the bridged metadata header is the same: They both pack each
field at a byte-aligned offset, then add padding to the next byte.