Enforce missing alignment constraint (#1315)
This PR detects the following case that is currently unhandled by action analysis: Suppose
ActionPhvConstraints detects an action writing to a container using ONE unallocated PHV source and
also action data/constant. In that case, the unallocated PHV source must be aligned at the same
position as its destination slice in the destination container.

After detecting this case, this PR generates the necessary alignment for the unallocated PHV source
using the existing code for generating conditional constraints. This new conditional constraint is
subsequently enforced by the existing code in AllocatePHV.