Remove duplicated bridged metadata in PHV alloc (#1314)
PHV allocation uses slice lists to (a) implement adjacency constraints, and (b)
to prevent fragmentation of constrained metadata.  Most field either fall in
category (a), (b), or none, but not both.  Bridged metadata is the exception,
and as a result it was being added to two slices, which complicates allocation
unnecessarily.

This commit prevents bridged metadata from being added to more than one slice
list.