Install ethtool in docker image
I tracked down the sporadic failures in the counter1 stf2ptf test to
send & receive offloads not being turned off for the veth interfaces in
the docker. Apparently, if we send packets too fast with PTF, the veth
kernel driver tries to do something funky with the packets, which leads
to packet drops. Turning off offloads solves the issue.

The interesting thing is that the veth_setup.sh script used for PTF
tests was supposed to turn off offloads, but because ethtool wasn't
installed in the docker, these commands were erroring silently. I seized
the opportunity to add "set -e" to the script...

On a side note, the packet drops seem to happen mostly when we send
arbitrary packets and not when we send regular IP/TCP traffic.