generate PHV logging classes from the phv schema (#1336)
schema-based logging generation

Adds a skeleton for structured logging based on a json schema, which
includes a script to generate C++ classes inheriting from
Logging::Logger.

The basic structure of the class generation is discussed in:
https://docs.google.com/document/d/1iwXXPAmk_LaGC8yXvScNrfutgYPBnyCFCQKh2vzpaKM

Classes for logging are generated from a json schema.  All classes are
generated in a Logging namespace.

The constructors take arguments for all basic types and provide append
methods for all the complex types (e.g. arrays and objects). Arrays are
represented as std::vector data members templetized with the appropriate
type. We use rapidjson to serialize the classes to a file.

A base Logging class is the main document. It is still work in
progress, the idea is to support logging levels, but not clear yet how
to partition the logging info based on the schema. It is likely that
we have to define the association of fields and levels in the schema
itself, and then serialize based on the level.

TODO: integrate the current classes with the PHV logging.

* enums, unions, and optional arguments

Optional arguments were not handled. Now we look into the required
arguments in the schema and order the constructor arguments
properly. We provide default values, such that if the arguments are
not set, they end up not being printed in the output .json. One issue
is with the enums, there is currently no "default" value. Might need
to add it.

Enums and unions also needed fixing and proper serialization to obtain
the values stored.

Also added destructors since the arrays store pointers to objects that
need deleting.

BRIG-462  #assign #start-work #fixed

* update JBay tests and xfails