Add PTF tests for fabric.p4 INT transit
Run PTF combination tests for fabric.p4

Since ONF uses different preprocessor flags for different switches in
the fabric. We leverage PTF's ability to classify tests in
group. Different preprocessor flags mean that different groups of tests
are run.

This required some minor changes in the runtest script, as well as in
p4c, which is why this commit also updates the submodule ref point for
p4c.

Compilation of fabric.p4 with INT transit enabled is currently failing
because a pending patch, which is why the tests are tagged as XFAIL.