Remove false warnings related to invalidate instruction (#1348)
When verifying whether any part of the container is incorrectly overwritten, action analysis uses
the phv and adi alignment objects to determine the total number of bits written in a container for a
particular action. However, for invalidate instructions, there are no reads and so, these alignment
objects cannot be used. This resulted in action analysis throwing false warnings about part of the
container being incorrectly overwritten. To address this, I have added special handling for
invalidate instructions, where the total number of bits written is calculated using the destination
slices in the container.