Introduce no-pack property for non-MOVE sources smaller than destination field size (#1352)
Currently, for non-move based operations, PHV allocation imposes a no_pack property on destinations
of the operations. However, the same no-pack property is not imposed on sources of those non-move
operations because of an implicit assumption that all operands will be the same bit width.

However, we do have cases such as the following:
    a = b + c, where a : 16b, b : 16b and c : 8b

In this case, we need to set the no_pack property for source c. In general, we are required to set
the no_pack property for the source which has a smaller width than the width of the destination.

BRIG-524 #assign #start-work #fixed