Action analysis handles cases with two unallocated PHV sources (#1384)
When action analysis encounters a packing with two unallocated PHV sources, it is necessary to
enforce alignment constraints on one of those PHV sources. This PR introduces those conditional
alignment constraints for the case where there are exactly two unallocated sources and no allocated
source/constant or action data source. In this case, the conditional alignment constraint is imposed
on the smaller of the two PHV sources to allow for the larger source slice to be placed in a number
of different starting positions within its container.

* Ignore destination slices (also used as source slices) when determining two unallocated sources

In case the destination slices are also used as a source slice, then ActionPhvConstraints should
treat them as already allocated (because the packing and position of destination slices for a call
of can_pack is fixed).

BRIG-515 #assign #start-work #fixed