adds support for running PD tests without p4factory (#1365)
adds support for running PD tests without p4factory

Enhances runtest and ptf_runner to run PD tests. There are several
additional dependencies, in particular it requires ptf-utils, which is
added as a submodule in p4-tests/p4testutils. (as a side note, since
we included this repo, we may want to consider synchronizing
base_tests.py with ptf-utlls/p4runtime_base_tests.py and get rid of
the former!). It also requires that the Docker image preserves build
tools (autoconf, automake, g++, libtool, at a minimum).

runtest gains one argument '-pd' that chooses to run pd tests.
In this case, it will:
 - invoke autogen.sh in the p4-build directory
 - configure the test with p4-build in the test.out directory
 - compile the P4 program and the PD libraries using p4-build
 - generate the driver configuration file
 - invokes ptf_runner with the appropriate arguments

Since PD compiles the P4 program as part of p4-build, there is no need
to invoke the compiler twice, so we skip the default compilation and
the check for the bfa output when compiled with PD. If the compilation
fails, we get an error anyway.

ptf_runnner gains the '--pdtest' argument that distinguishes between
p4runtime tests and pd tests. The main difference is the test_params
sent to PTF and the variety of python path to all the required
dependencies. Add the `--no-pi` option to start bf_switchd.

We add the `bfn_add_p4factory_tests` macro to TestUtils.cmake to add
all the p4factory tests as PTF pd tests. This replaces adding them as
compile only, and since there is still a thrift issue, we also add
them to xfail at the same time. This will need to be removed once we
fully get ptf_runner functional.

Adds cmake and runtest support to pass an individual test to
ptf_runner through a test specific environment variable.
If this succeeds, we need to consider two possible optimizations:
- extend ptf to accept more than one test; or
- extend ptf_runner to iterate through a bunch of these tests
and invoke ptf only for each of them.
This is done through an an environment variable: PTF_TEST_SPECS. 
It allows us to add arguments  to a test after the test has been "added", 
and we use this to select a subset ptf tests, otherwise the run would be too long.

We add a new script gen-pd-conf.py that takes the device.conf template
and specializes it for the test to run.

And that's about it, beside updating the refpoint to the p4-tests repo
to get the latest p4-build, since it has fixed the p4_name_lookup
issue. 

We add a selection of p4factory tests to Tofino. For JBay we restrict the
tests to just emulation for now. Quite a few of others do not compile,
we'll add them incrementally.

* add no-pi option to bf_switchd

And address reviewer's comments:
- add test-only option for PD tests
- add debug function and replace direct logger calls.
- update the README file