Fix mirror egress parser generation (#1406)
* Fix mirror egress parser generation

Recent changes to TNA changed the names related to the mirroring extern, eg.
"mirror_packet" to "Mirror", which breaks egress parser generation.  This
commit updates to use the right names.

The egress parser generation for mirror packets also needs a default case for
when (a) there are no field lists, or (b) there are field lists, but the mirror
session ID doesn't identify any of them.

Fixes BRIG-502.