Allocate CLOT fields to PHV when necessary (#1410)
Sometimes, CLOT fields need to be allocated to PHV containers, eg. to fill a
deparsed container when an adjacent field is added to a larger header field.

This takes a conservative, wholistic approach to CLOT candidate fields: When a
super cluster is composed entirely of CLOT candidates, remove the super cluster
entirely.  Otherwise, it must be the case that part of a set of CLOT-candidate
fields are adjacent to fields involved in MAU operations and may need to be
packed with them.

This also fixes a bug where CLOT fields were included in slice lists but were
excluded from allocation at the last minute, resulting in a header field not
completely filling a PHV container.