Reduce PHV/TPHV slice list splitting (#1347)
Remove the heuristic that breaks slice lists at PHV/TPHV boundaries, and the
one that eagerly splits slice lists into 32b chunks.  These aren't obviously
contributing to switch fitting, and they're preventing compilation of other
programs (fabric + INT), because PHV allocation doesn't recombine slice lists
once they've been split.

This commit also improves the PHV/TPHV split heuristic as follows, although
this is currently commented out:

As a heuristic when creating slice lists (adjacency constraints between header
fields), we start a new slice list when the previous field is a TPHV candidate
and the next isn't, or vice versa.

However, this splits a slice list too eagerly in the case when the list so far
already contains a mix of TPHV and PHV candidates, eg. when the fields are not
byte aligned.

This commit tunes the heuristic to only split when the entire slice list so far
is a TPHV candidate and the next field is a PHV candidate, or vice versa.

Fixes BRIG-518.

This PR also changes how PHV allocation deals with CLOT-allocated fields:

1. SuperClusters that entirely comprise CLOT-allocated fields are removed from PHV allocation before pre-slicing.

2. During PHV allocation, SuperClusters that entirely comprise CLOT-allocated slices (but not whole fields) are also omitted from allocation.

3. After PHV allocation, a P4C_UNIMPLEMENTED exception is thrown if any CLOT-allocated field has any slices allocated to PHV.