Initial work for generating BF-RT JSON file
We add the following:
 * a command-line flag (--bf-rt-schema) to p4c-barefoot and the p4c
 driver. This flag indicates that p4c-barefoot is required to generate
 the BF-RT JSON for the P4 program and to write it to the specified
 output file. The generation is done from the P4Info protobuf message,
 since most of the information is the same - although in a more uniform
 format -, which makes the code very straight-forward.
 * an executable Python script (gen_bf_rt_json_schema.py) which can
 generate the schema for the BF-RT JSON format and can be used to
 validate the output the --bf-rt-schema option for p4c-barefoot.

As of now, we only support a few table types: match-action (direct,
indirect, indirect with selector), action, selector, counter and
meter. In particular, we do not support any of the "fixed" tables
(e.g. PRE tables) yet. However, this is enough for the drivers team to
get started on the "P4 part" of BF-RT (which calls into pipe_mgr).

The schema for the BF-RT JSON is likely to change in the upcoming
months, but I think the drivers team will benefit from being able to use
the master branch of Brig.

A few things that need to be answered / improved:
 * common data field ids should probably be exposed to the BF-RT
 application as enums, which means we need a single source of truth for
 BF-RT and the compiler.
 * gen_bf_rt_json_schema.py should probably be called automatically by
 the p4c driver (in developper mode at least), just like we do for the
 context JSON.
 * is the schema sufficiently uniform? For example we could make the
 action data fields in the action spec even more similar to the common
 data fields, but would that make things easier on the drives side?